requirements.txt
asgiref==3.8.1
attrs==25.3.0
cffi==1.17.1
cryptography==45.0.3
dj-database-url==3.0.1
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.0
pytest-django==4.11.1
python-decouple==3.8
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
timedelta==2020.12.3
typing_extensions==4.14.0
uritemplate==4.2.0
whitenoise==6.11.0
